First Web Page
Making my first web page locally using NotePad++.  It will change a lot
the more I learn!  Considering I have only worked with Java and Android
applications it is a completely different experience.  To start, it is
much quicker to load your work compared to build times and emulators.
The code aspect is simple, and smooth.  It is nice having to design for
a web browser size but with time, I am sure it will be difficult to
format it to fit the mobile screen size.